Add evaluation of Bezier curves and surfaces
This implementation uses recursive templates to compute factorials and
binomial coefficients at compile time. It also uses recursive templates
to sum the terms of the finite mathematical series which forms the
Bezier function.

The Bernstein polynomials are computed using double precision but the
result has to be cast to the value_type of the vector type to ensure
that the vector type's multiplication function template can be applied.

It is required that the vector type provides multiplication and addition
functions, a length function, and an index operator.